Switch to local index for sparsity column

+++ cpp/dolfinx/la/SparsityPattern.h
@@ -101,7 +101,7 @@ public:
 
   /// Sparsity pattern for the un-owned (off-diagonal) columns. Uses global
   /// indices for the columns.
-  const graph::AdjacencyList<std::int64_t>& off_diagonal_pattern() const;
+  const graph::AdjacencyList<std::int32_t>& off_diagonal_pattern() const;
 
   /// Return MPI communicator
   MPI_Comm mpi_comm() const;
@@ -115,11 +115,11 @@ private:
 
   // Caches for diagonal and off-diagonal blocks
   std::vector<std::vector<std::int32_t>> _diagonal_cache;
-  std::vector<std::vector<std::int64_t>> _off_diagonal_cache;
+  std::vector<std::vector<std::int32_t>> _off_diagonal_cache;
 
   // Sparsity pattern data (computed once pattern is finalised)
   std::shared_ptr<graph::AdjacencyList<std::int32_t>> _diagonal;
-  std::shared_ptr<graph::AdjacencyList<std::int64_t>> _off_diagonal;
+  std::shared_ptr<graph::AdjacencyList<std::int32_t>> _off_diagonal;
 };
 } // namespace la
 } // namespace dolfinx
